+++ FE/src/components/Issue/Filter/index.tsx
@@ -1,0 +1,9 @@
+import * as S from './style';
+
+export default function Filter({ ...props }) {
+  return (
+    <S.FilterLayer>
+      <span>{props.label}</span>
+    </S.FilterLayer>
+  );
+}
