+++ FE/src/components/Issue/index.tsx
@@ -1,16 +1,9 @@
 import Navigation from '@components/Issue/Navigation';
 import Item from '@components/Issue/Item';
-import styled from '@emotion/styled';
 import { issueState } from '@recoil/atoms/issue';
 import { useQuery } from 'react-query';
 import { useRecoilState } from 'recoil';
-
-const IssueWrapperStyle = styled.section`
-  width: 80rem;
-  border: 1px solid;
-  border-radius: 1rem;
-  border-color: ${({ theme }) => theme.color.line};
-`;
+import { IssueWrapperLayer } from './style';
 
 export default function Issue() {
   const [issue, setIssue] = useRecoilState(issueState);
@@ -25,13 +18,13 @@ export default function Issue() {
 
   return (
     <>
-      <IssueWrapperStyle>
+      <IssueWrapperLayer>
         <Navigation />
         {issue &&
           issue.map(issueData => {
             return <Item issue={issueData} key={`issue-${issueData.id}`} />;
           })}
-      </IssueWrapperStyle>
+      </IssueWrapperLayer>
     </>
   );
 }
