+++ FE/src/components/Issue/style.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const IssueWrapperLayer = styled.section`
+  width: 80rem;
+  border: 1px solid;
+  border-radius: 1rem;
+  border-color: ${({ theme }) => theme.color.line};
+`;
